+++ packages/bloc_test/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 7.0.2
 
 - fix: `blocTest` timeouts when verify fails
+- fix: `blocTest` timeouts when expect fails
 
 # 7.0.1